Move passgen work up a level
Easiest to keep the files at the highest level, until I better
understand how go is generally organized.
