+++ encryption/passgen_test.go
@@ -48,5 +48,23 @@ func TestGenKeyfileUnique(t *testing.T) {
 	if !are_unique {
 		t.Fatal("Generated keyfiles match")
 	}
+}
 
+func TestGenPasswordLength(t *testing.T) {
+	for i := 1; i < 100; i++ {
+		password := GenPassword(i, true, true, true, true)
+		if len(password) != i {
+			t.Fatal("Incorrect password length")
+		}
+	}
+}
+
+func TestGenPasswordUnique(t *testing.T) {
+	for i := 0; i < 100; i++ {
+		passwordA := GenPassword(10, true, true, true, true)
+		passwordB := GenPassword(10, true, true, true, true)
+		if passwordA == passwordB {
+			t.Fatal("Randomly generated passwords match")
+		}
+	}
 }
